refactor: rename node implementations to *_node.py for clarity
