+++ src/copium_loop/nodes/coder_node.py
@@ -6,7 +6,7 @@ from copium_loop.state import AgentState
 from copium_loop.telemetry import get_telemetry
 
 
-async def coder(state: AgentState) -> dict:
+async def coder_node(state: AgentState) -> dict:
     telemetry = get_telemetry()
     telemetry.log_status("coder", "active")
     telemetry.log_output("coder", "--- Coder Node ---\n")
