+++ src/copium_loop/nodes/pr_creator_node.py
@@ -14,7 +14,7 @@ from copium_loop.state import AgentState
 from copium_loop.telemetry import get_telemetry
 
 
-async def pr_creator(state: AgentState) -> dict:
+async def pr_creator_node(state: AgentState) -> dict:
     telemetry = get_telemetry()
     telemetry.log_status("pr_creator", "active")
     telemetry.log_output("pr_creator", "--- PR Creator Node ---\n")
