+++ test/nodes/test_coder.py
@@ -7,7 +7,7 @@ from langchain_core.messages import HumanMessage, SystemMessage
 from copium_loop.nodes import coder
 
 # Get the module object explicitly to avoid shadowing issues
-coder_module = sys.modules["copium_loop.nodes.coder"]
+coder_module = sys.modules["copium_loop.nodes.coder_node"]
 
 
 @pytest.fixture
